+++ lib/features/cart/domain/repositories/cart_repository_interface.dart
@@ -4,7 +4,7 @@ import 'package:sixam_mart/interfaces/repository_interface.dart';
 abstract class CartRepositoryInterface<OnlineCart> extends RepositoryInterface<OnlineCart> {
   Future<void> addSharedPrefCartList(List<CartModel> cartProductList);
   @override
-  Future<dynamic> update(Map<String, dynamic> body, int? id, {double price, int quantity, bool isUpdateQty = false});
+  Future<dynamic> update(Map<String, dynamic> body, int? id, {double price, double quantity, bool isUpdateQty = false});
   @override
   Future<bool> delete(int? id, {bool isRemoveAll = false});
 }
